+++ kojiclient/apis/xmlrpc/xmlrpc_suite_test.go
@@ -4,7 +4,7 @@ import (
 	"errors"
 	"testing"
 
-	"github.com/odra/cowji/pkg/koji-client/meta"
+	"github.com/odra/cowji/kojiclient/meta"
 	. "github.com/onsi/ginkgo"
 	. "github.com/onsi/gomega"
 )
