+++ kojiclient/kojiclient.go
@@ -4,8 +4,8 @@ import (
 	"net/url"
 	"strings"
 
-	"github.com/odra/cowji/pkg/koji-client/apis/xmlrpc"
-	"github.com/odra/cowji/pkg/koji-client/meta"
+	"github.com/odra/cowji/kojiclient/apis/xmlrpc"
+	"github.com/odra/cowji/kojiclient/meta"
 )
 
 // A Koji API client struct, used to consume methods from Koji Hub.
